Make `LogicalPlan::with_new_exprs`, deprecate `from_plan`
